chore(types): move type definitions to index and global files
