+++ src/types/global.d.ts
@@ -1,0 +1,9 @@
+// global api ready flag and listener
+declare global {
+  interface Window {
+    onYouTubeIframeAPIReady: () => void
+    YT: any
+  }
+}
+
+export {}
